Add compilation test for built-in Greeter example contract
