+++ populus/assets/Greeter.sol
@@ -1,0 +1,17 @@
+pragma solidity ^0.4.0;
+
+contract Greeter {
+    string public greeting;
+
+    function Greeter() {
+        greeting = 'Hello';
+    }
+
+    function setGreeting(string _greeting) public {
+        greeting = _greeting;
+    }
+
+    function greet() constant returns (string) {
+        return greeting;
+    }
+}
