+++ populus/assets/test_greeter.py
@@ -1,0 +1,15 @@
+def test_greeter(chain):
+    greeter, _ = chain.provider.get_or_deploy_contract('Greeter')
+
+    greeting = greeter.call().greet()
+    assert greeting == 'Hello'
+
+
+def test_custom_greeting(chain):
+    greeter, _ = chain.provider.get_or_deploy_contract('Greeter')
+
+    set_txn_hash = greeter.transact().setGreeting('Guten Tag')
+    chain.wait.for_receipt(set_txn_hash)
+
+    greeting = greeter.call().greet()
+    assert greeting == 'Guten Tag'
