+++ tests/compilation/test_compiling_contract_sources.py
@@ -1,3 +1,7 @@
+import os
+
+from populus import ASSETS_DIR
+
 from populus.compilation import (
     compile_project_contracts,
 )
@@ -9,6 +13,8 @@ from populus.utils.testing import (
     load_contract_fixture,
     load_test_contract_fixture,
 )
+
+GREETER_SOURCE_PATH = os.path.join(ASSETS_DIR, 'Greeter.sol')
 
 
 @load_contract_fixture('Math.sol')
@@ -55,3 +61,10 @@ def test_compiling_with_abstract_contract_inhereted(project):
 
     assert 'Abstract' in contract_data
     assert 'UsesAbstract' in contract_data
+
+
+@load_contract_fixture(GREETER_SOURCE_PATH)
+def test_compiling_example_greeter_contract(project):
+    _, contract_data = compile_project_contracts(project)
+
+    assert 'Greeter' in contract_data
